Fix copy/move and use UUID correctly in frontend

+++ server/src/main/java/de/ingrid/igeserver/api/UploadApiController.kt
@@ -56,7 +56,7 @@ class UploadApiController @Autowired constructor(
         flowIdentifier: String,
         flowFilename: String,
     ): ResponseEntity<UploadResponse> {
-        val canWrite = aclService.getPermissionInfo(principal as Authentication, docId).canWrite
+        val canWrite = aclService.getPermissionInfo(principal as Authentication, docId.toInt()).canWrite
         if (!canWrite) {
             val uploadResponse = UploadResponse(ForbiddenException.withAccessRights("No access to referenced dataset"))
             return ResponseEntity<UploadResponse>(uploadResponse, HttpStatus.FORBIDDEN)
@@ -146,7 +146,7 @@ class UploadApiController @Autowired constructor(
         file: String,
         replace: Boolean
     ): ResponseEntity<UploadResponse> {
-        val canWrite = aclService.getPermissionInfo(principal as Authentication, docId).canWrite
+        val canWrite = aclService.getPermissionInfo(principal as Authentication, docId.toInt()).canWrite
         if (!canWrite) {
             val uploadResponse = UploadResponse(ForbiddenException.withAccessRights("No access to referenced dataset"))
             return ResponseEntity<UploadResponse>(uploadResponse, HttpStatus.FORBIDDEN)
@@ -171,7 +171,7 @@ class UploadApiController @Autowired constructor(
         val idx = requestURI.indexOf(docId)
         val file = URLDecoder.decode(requestURI.substring(idx + docId.length + 1), "UTF-8")
 
-        val canRead = aclService.getPermissionInfo(principal as Authentication, docId).canRead
+        val canRead = aclService.getPermissionInfo(principal as Authentication, docId.toInt()).canRead
         if (!canRead) {
             throw ForbiddenException.withAccessRights("No access to referenced dataset")
         }
@@ -198,7 +198,7 @@ class UploadApiController @Autowired constructor(
     }
 
     override fun deleteFile(principal: Principal, docId: String, file: String): ResponseEntity<Unit> {
-        val canWrite = aclService.getPermissionInfo(principal as Authentication, docId).canWrite
+        val canWrite = aclService.getPermissionInfo(principal as Authentication, docId.toInt()).canWrite
         if (!canWrite) {
             throw ForbiddenException.withAccessRights("No access to referenced dataset")
         }
